sec and graph tests for offset and raw pointer data structures
